# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Production Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  tongSLNgayHienTai: number;
  tongSLTuan: number;
  tongSLThang: number;
  tongPheNgay: number;
  tongPheTuan: number;
  tongPheThang: number;
  soLenhDangSX: number;
  tyLeHoanThanhTB: number;
  soRecordHomNay?: number;
  soRecordTuan?: number;
}