# lib/api-zod/src/generated/types/orderCompletion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Production Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderCompletion {
  donHang: string;
  maSoi: string;
  tenSoi: string;
  /** @nullable */
  slDonHangDaSX: number | null;
  /** @nullable */
  canSXTiep: number | null;
  /** @nullable */
  tyLeHoanThanh: number | null;
}