# lib/api-zod/src/generated/types/outputTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Production Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface OutputTrendPoint {
  date: string;
  tongSL: number;
  tongPhe: number;
  /** @nullable */
  slCa1?: number | null;
  /** @nullable */
  slCa2?: number | null;
  /** @nullable */
  slCa3?: number | null;
}