# lib/api-zod/src/generated/types/production.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Production Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Production {
  id: number;
  /** @nullable */
  stt?: number | null;
  ngayNhap: Date;
  ngaySanXuat: Date;
  /** @nullable */
  lenhXK?: string | null;
  /** @nullable */
  ngayBanHanhLXK?: string | null;
  maSoi: string;
  tenSoi: string;
  /** @nullable */
  tenSoiMoi?: string | null;
  donHang: string;
  /** @nullable */
  tyLeNhua?: number | null;
  /** @nullable */
  ngaySanXuatGanNhat?: string | null;
  /** @nullable */
  mayTSCa1?: string | null;
  /** @nullable */
  thoiGianTSCa1?: number | null;
  /** @nullable */
  sanLuongCa1?: number | null;
  /** @nullable */
  mayTSCa2?: string | null;
  /** @nullable */
  thoiGianTSCa2?: number | null;
  /** @nullable */
  sanLuongCa2?: number | null;
  /** @nullable */
  mayTSCa3?: string | null;
  /** @nullable */
  thoiGianTSCa3?: number | null;
  /** @nullable */
  sanLuongCa3?: number | null;
  /** @nullable */
  tongSLNgay?: number | null;
  /** @nullable */
  slDonHangDaSX?: number | null;
  /** @nullable */
  slLuyKeTuan?: number | null;
  /** @nullable */
  luyKeTHang?: number | null;
  /** @nullable */
  canSXTiep?: number | null;
  /** @nullable */
  dvt?: string | null;
  /** @nullable */
  phekeoMay?: number | null;
  /** @nullable */
  pheChayMay?: number | null;
  /** @nullable */
  pheChuyenDoi?: number | null;
  /** @nullable */
  pheSuCo?: number | null;
  /** @nullable */
  pheDungMay?: number | null;
  /** @nullable */
  pheXuLy?: number | null;
  /** @nullable */
  tongPheNgay?: number | null;
  /** @nullable */
  pheDonHang?: number | null;
  /** @nullable */
  pheLuyKeTuan?: number | null;
  /** @nullable */
  pheLuyKeTHang?: number | null;
  /** @nullable */
  tyLeHoanThanh?: number | null;
  /** @nullable */
  ghiChu?: string | null;
  createdAt?: string;
  updatedAt?: string;
}