# lib/api-zod/src/generated/types/productionListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Production Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Production } from "./production";

export interface ProductionListResponse {
  data: Production[];
  total: number;
  page: number;
  limit: number;
}