# lib/api-zod/src/generated/types/shiftPerformance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Production Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ShiftPerformance {
  ca1: number;
  ca2: number;
  ca3: number;
  ca1Pct?: number;
  ca2Pct?: number;
  ca3Pct?: number;
}