# lib/api-zod/src/generated/types/wasteBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Production Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface WasteBreakdown {
  phekeoMay: number;
  pheChayMay: number;
  pheChuyenDoi: number;
  pheSuCo: number;
  pheDungMay: number;
  pheXuLy: number;
}